+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 - Missing `DelayMs<u8>` / `DelayUs<u8>` impls for fugit::Delay
 - Support of embedded-hal 1.0.0-alpha.7 [#443]
+- `hd44780` example [#441]
 - Aliases for peripheral wrappers [#434]
 - `WithPwm` trait implemented for timers with channels (internals) [#425]
 - `Pwm` struct with `split` method and implementation of embedded-hal::Pwm (similar to f1xx-hal) [#425]
@@ -52,6 +53,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 [#439]: https://github.com/stm32-rs/stm32f4xx-hal/pull/439
 [#440]: https://github.com/stm32-rs/stm32f4xx-hal/pull/440
 [#443]: https://github.com/stm32-rs/stm32f4xx-hal/pull/443
+[#441]: https://github.com/stm32-rs/stm32f4xx-hal/pull/441
 
 ### Changed